refactor(protocol): consolidate keycode protocol scoping helpers
